Force EnvisalinkAlarm._attr_code_arm_required=false
At least on my DSC system, it is not required to have a code for arming the alarm; only do disarm.
It does what I want if I set this attribute to always false.

Since I do not want HomeAssistant to be able to disarm my alarm without having to enter a user code,
But I want it to be able to arm it, having always false works for me, but maybe we need to have this
as an extra configuration option or only for my DSC device?

Signed-off-by: David Racine <bass_dr@hotmail.com>